VecchiaMLE defines

VecchiaMLE.SPARSITY_GEN — Constant

Specification for the Sparsity Pattern generation algorithm.

Supported Sparsity Patterns

NearestNeighbors (:NN) : Based on https://github.com/KristofferC/NearestNeighbors.jl

HNSW (:HNSW) : Based on https://github.com/JuliaNeighbors/HNSW.jl

Custom (:USERGIVEN) : Given by user. You don't need to fill this field out, but the given sparsity pattern must be in CSC format.

VecchiaMLE.VecchiaCache — Type

Internal struct from which to fetch persisting objects in the optimization function. There is no need for a user to mess with this!

Fields

n : Size of the problem

M : Number of Samples

nnzL : Number of nonzeros in L

colptrL : Array which indexes the beginning of each column in L

rowsL : Row index of nonzero entries in L

diagL : Position of the diagonal coefficient of L

m : Number of nonzeros in each column of L

offsets : Number of nonzeros in hessobjvals before the block Bⱼ

B : Vector of matrices Bⱼ, the constant blocks in the Hessian

nnzh_tri_obj : Number of nonzeros in the lower triangular part of the Hessian of the objective

nnzh_tri_lag : Number of nonzeros in the lower triangular part of the Hessian of the Lagrangian

hess_obj_vals : Nonzeros of the lower triangular part of the Hessian of the objective

buffer : Additional buffer needed for the objective

VecchiaMLE.VecchiaModel — Type

Struct needed for NLPModels. There is no need for the user to mess with this!

VecchiaMLE.VecchiaMLEInput — Type

Input to the VecchiaMLE analysis.

Fields

k::Int: Number of neighbors, representing the number of conditioning points in the Vecchia Approximation.

samples::M: Samples to generate the output. Each sample should match the length of the observed_pts vector.

Keyword Arguments

ptset::AbstractVector # The locations of the analysis. May be passed as a matrix or vector of vectors.

rowsL::AbstractVector # The sparsity pattern rows of L if the user gives one. MUST BE IN CSC FORMAT!

colptrL::AbstractVector # The column pointer of L if the user gives one. MUST BE IN CSC FORMAT!

metric::Distances.metric # The metric by which nearest neighbors are determined. Defaults to Euclidean.

sparsitygen::Symbol # The method by which to generate a sparsity pattern. See SPARSITY_GEN.

VecchiaMLE utils

VecchiaMLE.generate_samples — Function

Samples_Matrix = generate_samples(MatCov::AbstractMatrix, 
                                  number_of_samples::Int;
                                  arch::Symbol)

Generate a number of samples according to the given Covariance Matrix MatCov.
Note the samples are given as mean zero.

Input arguments

MatCov: A Covariance Matrix, assumed positive definite;

n: The length of one side of the Covariance matrix;

number_of_samples: How many samples to return.

Keyword Arguments

arch: Either to perform the linear algebra on cpu or gpu. See ARCHITECTURES. Defaults to :cpu.

Output arguments

Samples_Matrix : A matrix of size (numberofsamples × n), where the rows are the i.i.d samples

VecchiaMLE.generate_MatCov — Function

Covariance_Matrix = generate_MatCov(params::AbstractArray,
                                    ptset::AbstractVector)

Generates a matern-like Covariance Matrix determined via the given paramters (params) and ptset. 
See https://en.wikipedia.org/wiki/Mat%C3%A9rn_covariance_function.

Input arguments

params: An array of length 3 that holds the parameters to the matern covariance kernel (σ, ρ, ν);

ptset: A set of points in 2D space upon which we determine the indices of the Covariance matrix;

Output arguments

Covariance_Matrix : An n × n Symmetric Matern matrix, where n = length(ptset).

Covariance_Matrix = generate_MatCov(n::Int, 
                                    params::AbstractArray)

Generates a 2D matern-like Covariance Matrix determined via the given paramters (params). 
A square grid with `n` elements will be generated and used in lieu of a given location set.
See https://en.wikipedia.org/wiki/Mat%C3%A9rn_covariance_function.

Input arguments

n: The dimension of the desired MatCov;

params: An array of length 3 that holds the parameters to the matern covariance kernel (σ, ρ, ν);

Output arguments

Covariance_Matrix : An n × n Symmetric Matern matrix

Covariance_Matrix = generate_MatCov(params::AbstractArray,
                                    ptset::AbstractMatrix)

Generates a matern-like Covariance Matrix determined via the given paramters (params). 
The matrix of locations will be parsed as a vector of vectors, where each location is a row vector. 
See https://en.wikipedia.org/wiki/Mat%C3%A9rn_covariance_function.

Input arguments

params: An array of length 3 that holds the parameters to the matern covariance kernel (σ, ρ, ν);

ptset : A matrix filled with row vectors of locations.

Output arguments

Covariance_Matrix : An n × n Symmetric Matern matrix, where n = size(ptset, 1).

VecchiaMLE.generate_xyGrid — Function

xyGrid = generate_xyGrid(n::Int)

A helper function to generate a point grid which partitions the positive unit square [0, 1] × [0, 1].
NOTE: This function at larger n values (n > 100) causes ill conditioning of the genreated Covariance matrix!
See generate_safe_xyGrid() for higher n values.

Input arguments

n: The length of the desired xyGrid;

Output arguments

xyGrid : The desired points in 2D space

VecchiaMLE.generate_safe_xyGrid — Function

xyGrid = generate_safe_xyGrid(n::Int)

A helper function to generate a point grid which partitions the positive square [0, k] × [0, k], 
where k = cld(n, 10). This should avoid ill conditioning of a Covariance Matrix generated by these points.

Input arguments

n: The length of the desired xyGrid;

Output arguments

xyGrid : The desired points in 2D space

VecchiaMLE.generate_rectGrid — Function

rect = generate_rectGrid(dims::Tuple)

A helper function to generate a point grid which partitions the positive square [0, 1] × [0, 1] by dims[1] and dims[2].

Input arguments

dims: A tuple of dimensions (nx, ny);

Output arguments

rectGrid : The desired points in 2D space as a vector of 2-element arrays

VecchiaMLE internals

VecchiaMLE.nn_to_csc — Function

rows, colptr = nn_to_csc(sparmat::Matrix{Float64})

A helper function to generate the sparsity pattern of the vecchia approximation (inverse cholesky) based 
on each point's nearest neighbors. If there are n points, each with k nearest neighbors, then the matrix
sparmat should be of size n x k. 

NOTE: The Nearest Neighbors algorithm should only consider points which appear before the given point. If
you do standard nearest neighbors and hack off the indices greater than the row number, it will not work. 

TODO: Can be parallelized. GPU kernel?

Input arguments

sparmat: The n x k matrix which for each row holds the indices of the nearest neighbors in the ptset.

Output arguments

rows: A vector of row indices of the sparsity pattern for L, in CSC format.

colptr: A vector of incides which determine where new columns start.

VecchiaMLE permutations

VecchiaMLE.IndexReorder — Function

permutation, dists = IndexReorder(condset::AbstractVector,
                                   data::AbstractVector,
                                   mu::AbstractVector,
                                   method::String = "standard", 
                                   reverse_ordering::Bool = true)

Front End function to determine the reordering of the given indices (data).
At the moment, only random, and maxmin are implemented.    

I BELIEVE the intention is to permute the coordinates of the samples with it, e.g.,
samples = samples[:, permutation]. Note for small conditioning sets per point (extremely sparse L),
this is not worth while!.

Input arguments

condset: A set of points from which to neglect from the index permutation;

data: The data to determine said permutation;

mu: The theoretical mean of the data, not mean(data)!

method: Either permuting based on maxmin, or random.

reverse_ordering: permutation is generated either in reverse or not.

Output arguments

permutation: The index set which permutes the given data.

dists: The array of maximum distances to iterating conditioning sets.

VecchiaMLE.IndexMaxMin — Function

See IndexReorder().

VecchiaMLE.IndexRandom — Function

See IndexReorder().

VecchiaMLE.dist_from_set — Function

Helper function. See IndexReorder().

VecchiaMLE sparsity patterns

VecchiaMLE.sparsity_pattern — Function

rows, colptr = sparsity_pattern(
    ::Val{<:Symbol}, 
    data::AbstractVector,
    k::Int,
    metric::Distances.Metric=Distances.Euclidean()
)

Generates the sparsity pattern of the approximate 
precision's cholesky factor, L, in CSC format. The pattern is 
determined by nearest neighbors of the previous points in data.

Input arguments

data: The point grid that was used to either generate the covariance matrix, or any custom ptset.

k: The number of nearest neighbors for each point (Including the point itself)

method: How the sparsity pattern is generated. See SPARSITY_GEN

metric: The metric for determining nearest neighbors.

Output arguments

rows: A vector of row indices of the sparsity pattern for L, in CSC format.

colptr: A vector of incides which determine where new columns start.

VecchiaMLE.sparsity_pattern_HNSW — Function

See sparsity_pattern(). In place for HNSW.jl

VecchiaMLE.sparsity_pattern_NN — Function

See sparsity_pattern(). Uses NearestNeighbors library. In case of tie, opt for larger index.