VecchiaMLE Defines

VecchiaMLE.COMPUTE_MODE — Type

Computation mode for which the analysis to run. Generally, we should see better performance at higher n values for the GPU.

VecchiaMLE.PRINT_LEVEL — Type

Print level of the program. Not implemented yet, but will be given by the user to determine the print level of both VecchiaMLE and MadNLP.

VecchiaMLE.ConfigManager — Type

At the moment, not used!

VecchiaMLE.VecchiaCache — Type

Internal struct from which to fetch persisting objects in the optimization function. There is no need for the user to mess with this!

VecchiaMLE.Diagnostics — Type

The Diagnostics struct that records some internals of the program that would be otherwise difficult to recover. The fields to the struct are as follows:

create_model_time: Time taken to create Vecchia Cache and MadNLP init.

LinAlg_solve_time: Time in LinearAlgebra routines in MadNLP

solve_model_time: Time taken to solve model in MadNLP

objective_value: Optimal Objective value.

normed_constraint_value: Optimal norm of constraint vector.

normed_grad_value: Optimal norm of gradient vector.

MadNLP_iterations: Iterations for MadNLP to reach optimal.

mode: Operation mode: CPU or GPU

VecchiaMLE.VecchiaModel — Type

Struct needed for NLPModels. There is no need for the user to mess with this!

VecchiaMLE.VecchiaMLEInput — Type

Input to the VecchiaMLE analysis, needs to be filled out by the user! The fields to the struct are as follows:

n: The Square root size of the problem. I.e., the length of one side of ptGrid.

k: The "number of neighbors", number of conditioning points regarding the Vecchia Approximation.

samples: The Samples in which to generate the output. If you have no samples consult the Documentation.

Number_of_Samples: The Number of Samples the user gives input to the program.

MadNLP_print_level: The print level to the optimizer. Can be ignored, and will default to ERROR.

mode: The opertaing mode to the analysis(GPU or CPU). The mapping is [1: 'CPU', 2: 'GPU'].

VecchiaMLE input

VecchiaMLE.VecchiaMLE_Run — Function

diagnostics, result = VecchiaMLE_Run(iVecchiaMLE::VecchiaMLEInput)

Main function to run the analysis. Make sure that your VecchiaMLEInput struct is 
filled out correctly!

Input arguments

iVecchiaMLE: The filled out VecchiaMLEInput struct. See VecchiaMLEInput struct for more details.

Output arguments

diagnostics: Some diagnostics of the analysis that would be difficult otherwise to get. See Diagnostics struct for more details.

pres_chol: The approximate precision's cholesky, generated by the samples.

VecchiaMLE.VecchiaMLE_Run_Analysis! — Function

See VecchiaMLE_Run().

VecchiaMLE.ExecuteModel! — Function

See VecchiaMLE_Run().

VecchiaMLE utils

VecchiaMLE.covariance2D — Function

Covariance_Matrix = covariance2D(xyGrid::AbstractVector, 
                                 params::AbstractVector)

Generate a Matern-Like Covariance Matrix for the parameters and locations given.
Note: This should not be called by the user. This is the back end for the function
generate_MatCov()!

Input arguments

xyGrid: A set of points in 2D space upon which we determine the indices of the Covariance matrix;

params: An array of length 3 (or 4) that holds the parameters to the matern covariance kernel (σ, ρ, ν).

Output arguments

Covariance_Matrix : An n × n Matern matrix, where n is the length of the xyGrid

VecchiaMLE.get_vecchia_model — Function

model = get_vecchia_model(iVecchiaMLE::VecchiaMLEInput, ptGrid::AbstractVector)

creates and returns a vecchia model based on the VecchiaMLEInput and point grid.

Input arguments

iVecchiaMLE: The filled out VecchiaMLEInput struct

ptGrid: The grid of locations

Output arguments

model: The Vecchia model based on the VecchiaMLEInput

VecchiaMLE.generate_Samples — Function

Samples_Matrix = generate_Samples(MatCov::AbstractMatrix, 
                                  n::Integer,
                                  Number_of_Samples::Int;
                                  mode::VecchiaMLE.COMPUTE_MODE )

Generate a number of samples according to the given Covariance Matrix MatCov.
Note the samples are given as mean zero. 
If a CUDA compatible device is detected, the samples are generated on the GPU
and transferred back to the CPU.

Input arguments

MatCov: A Covariance Matrix, presumably positive definite;

n: The length of one side of the Covariance matrix;

Number_of_Samples: How many samples to return.

Keyword Arguments

mode: Either generate samples on GPU or CPU

Output arguments

Samples_Matrix : A matrix of size (NumberofSamples × n), where the rows are the i.i.d samples

VecchiaMLE.generate_MatCov — Function

Covariance_Matrix = generate_MatCov(n::Integer,
                                    params::AbstractArray,
                                    ptGrid::AbstractVector)

Generates a Matern Covariance Matrix determined via the given paramters (params) and ptGrid.

Input arguments

n: The length of the ptGrid;

params: An array of length 3 (or 4) that holds the parameters to the matern covariance kernel (σ, ρ, ν);

ptGrid: A set of points in 2D space upon which we determine the indices of the Covariance matrix;

Output arguments

Covariance_Matrix : An n × n Symmetric Matern matrix

VecchiaMLE.generate_xyGrid — Function

xyGrid = generate_xyGrid(n::Integer)

A helper function to generate a point grid which partitions the positive unit square [0, 1] × [0, 1].
NOTE: This function at larger n values (n > 100) causes ill conditioning of the genreated Covariance matrix!
See generate_safe_xyGrid() for higher n values.

Input arguments

n: The length of the desired xyGrid;

Output arguments

xyGrid : The desired points in 2D space

VecchiaMLE.generate_safe_xyGrid — Function

xyGrid = generate_safe_xyGrid(n::Integer)

A helper function to generate a point grid which partitions the positive square [0, k] × [0, k], 
where k = cld(n, 10). This should avoid ill conditioning of a Covariance Matrix generated by these points.

Input arguments

n: The length of the desired xyGrid;

Output arguments

xyGrid : The desired points in 2D space

VecchiaMLE.MadNLP_Print_Level — Function

log_level = MadNLP_Print_Level(pLEvel::Integer)

A helper function to convert an integer to a MadNLP LogLevel.
The mapping is [1: 'TRACE', 2: 'DEBUG', 3: 'INFO', 4: 'WARN', 5: 'ERROR'].
Any other integer given is converted to MadNLP.Fatal.

Input arguments

pLevel: The given log level as an integer;

Output arguments

log_level : The coded MadNLP.LogLevel.

VecchiaMLE.Int_to_Mode — Function

cpu_mode = Int_to_Mode(n::Integer)

A helper function to convert an integer to a COMPUTE_MODE.
The mapping is [1: 'CPU', 2: 'GPU'].
Any other integer given is converted CPU.

Input arguments

n: The given cpu mode as an integer;

Output arguments

cpu_mode : The coded COMPUTE_MODE.

VecchiaMLE.Uni_Error — Function

Error = Uni_Error(TCov::AbstractMatrix,
                  L::AbstractMatrix)
Generates the "Univariate KL Divergence" from the Given True Covariane matrix 
and Approximate Covariance matrix. The output is the result of the following 
optimization problem: sup(a^T X_A || a^⊺ X_T). The solution is 
            f(μ) = ln(μ) + (2μ)^{-2} - 0.5,
where μ is the largest or smallest eigenvalue of the matrix Σ_A^{-1}Σ_T, whichever
maximizes the function f(μ). Note: Σ_A , Σ_T are the respective approximate and true
covariance matrices, and X_A, X_T are samples from the respective Distributions (mean zero).

Input arguments

TCov: The True Covariance Matrix: the one we are approximating;

L: The cholesky factor of the approximate precision matrix.

Output arguments

f(μ) : the result of the function f(μ) detailed above.

VecchiaMLE.IndexReorder — Function

indx_perm, dist_set = IndexReorder(Cond_set::AbstractVector,
                                   data::AbstractVector,
                                   mean_mu::AbstractVector,
                                   method::String = "random", 
                                   reverse_ordering::Bool = true)

Front End function to determine the reordering of the given indices (data).
At the moment, only random, and maxmin are implemented.    

I BELIEVE the intention is to permute the coordinates of the samples with it, e.g.,
samples = samples[:, indx_perm]. Note for small conditioning sets per point (extremely sparse L),
this is not worth while!.

Input arguments

Cond_set: A set of points from which to neglect from the index permutation;

data: The data to determine said permutation;

mean_mu: The theoretical mean of the data, not mean(data)!

method: Either permuting based on maxmin, or random.

reverse_ordering: permutation is generated either in reverse or not.

Output arguments

indx_perm: The index set which permutes the given data.

dist_set: The array of maximum distances to iterating conditioning sets.

VecchiaMLE.IndexMaxMin — Function

See IndexReorder().

VecchiaMLE.dist_from_set — Function

Helper function. See IndexReorder().

VecchiaMLE.IndexRandom — Function

See IndexReorder().

VecchiaMLE.KLDivergence — Function

KL_Divergence = KLDivergence(TCov::Symmetric{Float64},
                             AL::AbstractMatrix)

Computes the KL Divergence of the True Covariance matrix, TCov, and
The APPROXIMATE INVERSE CHOLESKY FACTOR, AL. Assumed mean zero.
Note: This is extremely slow!

Input arguments

TCov: The True Covariance Matrix;

AL: The cholesky factor to the approximate precision matrix. I.e., The ouptut of VecchiaMLE.

Output arguments

res: The result of the KL Divergence function.

VecchiaMLE.SparsityPattern — Function

rows, cols, colptr = SparsityPattern(data::AbstractVector,
                                     k::Integer,
                                     format::String = "")

Front end to generate the sparsity pattern of the approximate 
precision's cholesky factor, L, in CSC format. The pattern is 
determined by nearest neighbors of the previous points in data.

Input arguments

data: The point grid that was used to either generate the covariance matrix, or any custom ptGrid.

k: The number of nearest neighbors for each point (Including the point itself)

format: Either blank or "CSC". This can be ignored.

Output arguments

rows: A vector of row indices of the sparsity pattern for L, in CSC format.

cols: A vector of column indices of the sparsity pattern for L, in CSC format.

colptr: A vector of incides which determine where new columns start.

VecchiaMLE.SparsityPattern_Block — Function

See SparsityPattern().

VecchiaMLE.SparsityPattern_CSC — Function

See SparsityPattern().

VecchiaMLE.sanitize_input! — Function

sanitize_input!(iVecchiaMLE::VecchiaMLEInput)

A helper function to catch any inconsistencies in the input given by the user.

The current checks are:

* Ensuring n > 0.
* Ensuring k <= n^2 (Makes sense considering the ptGrid and SparsityPattern sizes).
* Ensuring the sample matrix, if the user gives one, is nonempty and is the same size as n^2.
* Ensuring the MadNLP_print_level in 1:5. See MadNLP_Print_Level().
* Ensuring the mode in 1:2. See Int_to_Mode().

Input arguments

iVecchiaMLE: The filled-out VecchiaMLEInput struct. See VecchiaMLEInput struct for more details.